Conduct the cross wavelet transform on a time series or signal

Compute the cross wavelet transform using a complex Morlet wavelet based on the "analyze.coherency" function of the WaveletComp package.

analyze_xwavelet(
  data_1,
  data_2,
  dj = 1/100,
  lowerPeriod = 2,
  upperPeriod = 1024,
  verbose = FALSE,
  omega_nr = 8
)

Arguments

data_1

First input data set as a matrix or data frame. The first column must represent depth or time, and the second column the signal or proxy record.

data_2

Second input data as a matrix or data frame. The first column must represent depth or time, and the second column the signal or proxy record.

lowerPeriod

Minimum period to be analysed. Controls the highest analysed frequency.

upperPeriod

Maximum period to be analysed. Controls the lowest analysed frequency.

verbose

Logical. If TRUE, print progress and interpolation information.

Value

A list with class "analyze.Xwavelet" containing Wave: complex wavelet coefficients Power: time frequency power spectrum dt: sampling interval after interpolation Phase: instantaneous phase of the signal dj: number of frequencies Power.avg: average spectral power Period: physical periods corresponding to frequencies nc: number of columns in the input data nr: number of frequency levels axis.1: x axis values (time or depth) axis.2: y axis values (log2 scaled periods) c1: base number of wavelet cycles o: numeric vector of length two defining the minimum and maximum superlet order. If NULL, all frequencies are analysed using order one. x1: interpolated x values first data set y1: interpolated signal values first data set x2: interpolated x values second data set y2: interpolated signal values second data set

References

Moca, V. V., Bârzan, H., Nagy-Dăbâcan, A., and Mureșan, R. C. (2021). Time frequency super resolution with superlets. Nature Communications, 12, 337. https://doi.org/10.1038/s41467-020-20539-9

Author

The "analyze_Xsuperlet" that generates the input for the plotting function is based on the matlab code in Moca et al. (2021) and the the "analyze.coherency" function of the 'WaveletComp' R package

Examples

#Example 1. A cross superlet of two etp solutions with noise overprint
etp_1 <- etp(
 tmin = 0,
 tmax = 1500,
 dt = 1,
 eWt = 1.5,
 oWt = 0.75,
 pWt = 1,
 esinw = TRUE,
 standardize = TRUE,
 genplot = FALSE,
 verbose = FALSE
)
#> Error in etp(tmin = 0, tmax = 1500, dt = 1, eWt = 1.5, oWt = 0.75, pWt = 1,     esinw = TRUE, standardize = TRUE, genplot = FALSE, verbose = FALSE): could not find function "etp"

etp_2 <- etp(
 tmin = 0,
 tmax = 1500,
 dt = 1,
 eWt = 1,
 oWt = 0.5,
 pWt = 1.5,
 esinw = TRUE,
 standardize = TRUE,
 genplot = FALSE,
 verbose = FALSE
)
#> Error in etp(tmin = 0, tmax = 1500, dt = 1, eWt = 1, oWt = 0.5, pWt = 1.5,     esinw = TRUE, standardize = TRUE, genplot = FALSE, verbose = FALSE): could not find function "etp"

etp_1[, 2] <- etp_1[, 2] + colorednoise::colored_noise(
 nrow(etp_1),
 sd = sd(etp_1[, 2]) / 1.5,
 mean = mean(etp_1[, 2]),
 phi = 0.9
)
#> Error: object 'etp_1' not found
etp_2[, 2] <- etp_2[, 2] + colorednoise::colored_noise(
 nrow(etp_2),
 sd = sd(etp_2[, 2]) / 1.5,
 mean = mean(etp_2[, 2]),
 phi = 0.9
)
#> Error: object 'etp_2' not found

Xetp <- analyze_xwavelet(
 data_1 = etp_1,
 data_2  = etp_2,
 upperPeriod = 1024,
 lowerPeriod = 2,
 verbose = FALSE
)
#> Error: object 'etp_1' not found