Windowed timeOpt sedimentation rate estimation

The win_timeOpt function for conducts a widowed timeOpt sedimentation rate estimation This function is based on the eTimeOpt but allows for multithreaded analysis speeding up the process of conducting a Windowed timeOpt sedimentation rate estimation

win_timeOpt(
  data = NULL,
  window_size = 10,
  sedmin = 0.5,
  sedmax = 2,
  numsed = 100,
  limit = FALSE,
  fit = 2,
  fitModPwr = TRUE,
  flow = NULL,
  fhigh = NULL,
  roll = 10^6,
  targetE = c(405.7, 130.7, 123.8, 98.9, 94.9),
  targetP = c(20.9, 19.9, 17.1, 17.2),
  detrend = TRUE,
  normalize = TRUE,
  linLog = 1,
  run_multicore = FALSE,
  verbose = FALSE
)

Arguments

data

Input data set should consist of a matrix with 2 columns with the first column being depth and the second column being a proxy Default=NULL

window_size

size of the moving window in metres Default=15

sedmin

Minimum sedimentation rate for investigation (cm/ka). Default=0.1

sedmax

Maximum sedimentation rate for investigation (cm/ka). Default=1

numsed

Number of sedimentation rates to investigate in optimization grid. Default=100

limit

Limit evaluated sedimentation rates to region in which full target signal can be recovered? .Default=FALSE

fit

Test for (1) precession amplitude modulation or (2) short eccentricity amplitude modulation? Default=2

fitModPwr

Include the modulation periods in the spectral fit? Default=TRUE

flow

Low frequency cut-off for Taner bandpass (half power point in cycles/ka) Default=TRUE

fhigh

High frequency cut-off for Taner bandpass (half power point; in cycles/ka) Default=NULL

roll

Taner filter roll-off rate, in dB/octave. Default=c(10^6)

targetE

A vector of eccentricity periods to evaluate (in ka). These must be in order of decreasing period, with a first value of 405 ka. Default= "c(405.7, 130.7, 123.8, 98.9, 94.9)"

targetP

A vector of precession periods to evaluate (in ka). These must be in order of decreasing period. Default=c(20.9, 19.9, 17.1, 17.2)

detrend

Remove linear trend from data series? Default=TRUE

normalize

normalize the r2 curves of individual timeOpt runs Default=TRUE

linLog

Use linear or logarithmic scaling for sedimentation rate grid spacing? (0=linear, 1=log; default value is 1) Default=1

run_multicore

Run function using multiple cores Default=FALSE

verbose

print text Default=FALSE

Value

Returns a list which contains 10 elements element 1: r_2_envelope matrix element 2: r_2_power matrix element 3: r_2_opt matrix element 4: r_2_envelope_avg element 5: r_2_opt_avg element 6: depth element 7: y_axis element 8: linLog value

References

Routines for astrochronologic testing, astronomical time scale construction, and time series analysis <doi:10.1016/j.earscirev.2018.11.015>

Author

Based on the eTimeOpt function of the 'astrochron' R package.

Examples

# \donttest{
#Conduct a windowed timeOpt on the magnetic susceptibility record
#of the Sullivan core of Pas et al., (2018).
mag_win_timeOpt <-win_timeOpt(
data = mag,
window_size = 15,
sedmin = 0.1,
sedmax = 1,
numsed = 100,
limit = FALSE,
fit = 2,
fitModPwr = TRUE,
flow = NULL,
fhigh = NULL,
roll = 10 ^ 6,
targetE = c(405.7, 130.7, 123.8, 98.9, 94.9),
targetP = c(20.9, 19.9, 17.1, 17.2),
detrend = TRUE,
normalize =TRUE,
linLog = 1,
run_multicore =FALSE,
verbose=FALSE)
# }