Generate a log2 spaced EHA spectra

Compute a log2 spaced Evolutive Harmonic Analysis (EHA) & Evolutive Power Spectral Analysis This is a wrapper function for the "eha" function of the 'astrochron' R package

eha_log2(
  data = NULL,
  win = NULL,
  tbw = NULL,
  demean = NULL,
  detrend = NULL,
  upperPeriod = NULL,
  lowerPeriod = NULL,
  pad = NULL,
  padding = "noise"
)

Arguments

data

Input data, should be a matrix or data frame in which the first column is depth or time and the second column is proxy record.

win

Window size for EHA, in units of space or time.

tbw

MTM time-bandwidth product (<=10)

demean

Remove mean from data series? (T or F)

detrend

Remove linear trend from data series? (T or F)

upperPeriod

Upper period to be analyzed. The CWT analyses the signal starting from the lowerPeriod to the upperPeriod so the proper selection these parameters allows to analyze the signal for a specific range of cycles. scaling is done using power 2 so for the best plotting results select a value to the power or 2.

lowerPeriod

Lowest period to be analyzed. The CWT analyses the signal starting from the lowerPeriod to the upperPeriod so the proper selection these parameters allows to analyze the signal for a specific range of cycles. scaling is done using power 2 so for the best plotting results select a value to the power or 2.

padding

pad the edges of the data set with half a window length with the following, the "Mean", "noise" or "zero"

Pad

with zeros to how many points? Must not factor into a prime number >23. Maximum number of points is 200,000.

Value

The output is a list (analyze.eha object) which contain 12 objects which are the result of running the eha. Object 1: depth - depth/time of axis Object 2: log2_period - log2 scales period Object 3: pwr - Power values of the EHA run Object 4: amp - Amplitude values of the EHA run Object 5: prob - Probability values of the EHA run Object 6: f_test - F test values of the EHA run Scale size Object 7: Power.avg - Average power values Object 8: amp.avg - Average amplitude values Object 9: prob.avg - Average probability value Object 10: f_test.avg - Average f test values Object 11: x - x values of the data set Object 12: y - y values of the data set Object 13: window size

References

S.R. Meyers, 2012, Seeing Red in Cyclic Stratigraphy: Spectral Noise Estimation for #'Astrochronology: Paleoceanography, 27, PA3228, <doi:10.1029/2012PA002307>

S.R. Meyers, 2019 Cyclostratigraphy and the problem of astrochronologic testing, Earth-Science Reviews, Volume 190, <doi.org/10.1016/j.earscirev.2018.11.015.>

Author

Code based on on the "eha" function of the 'astrochron' R package

Examples

# \donttest{
#Example 1. A plot of a wavelet spectra using the Total Solar Irradiance
# data set of Steinhilber et al., (2012)

TSI_eha_log2 <- eha_log2(data = TSI,
win =  8192,
tbw = 4,
demean = TRUE,
detrend = TRUE,
upperPeriod = 8192,
lowerPeriod = c,
pad = NULL,
padding = "noise")
#> Error in eha(data_padded, tbw = tbw, pad = pad * 5, fmin = 1/upperPeriod,     fmax = 1/lowerPeriod, win = win, demean = T, detrend = T,     siglevel = 0.9, sigID = F, ydir = 1, output = 1, pl = 1,     palette = 6, centerZero = T, ncolors = 100, genplot = 0,     verbose = FALSE): could not find function "eha"

#Example 2. A plot of a wavelet spectra using the magnetic susceptibility
#data set of Pas et al., (2018)
mag_eha_log2 <- eha_log2(data = mag,
win =  254,
tbw = 4,
demean = TRUE,
detrend = TRUE,
upperPeriod = 254,
lowerPeriod = 0.1,
pad = NULL,
padding = "noise")
#> Error in xj[i]: only 0's may be mixed with negative subscripts

#Example 3. A plot of a wavelet spectra using the greyscale
# data set of Zeeden et al., (2013)
grey_eha_log2 <- eha_log2(data = grey,
win =  256,
tbw = 4,
demean = TRUE,
detrend = TRUE,
upperPeriod = 256,
lowerPeriod = 0.02,
pad = NULL,
padding = "noise")
#> Error in xj[i]: only 0's may be mixed with negative subscripts

# }