exponentially ramped sedimentation rate model to convert time to stratigraphy

Apply a exponentially increasing increasing sedimentation rate model to convert time to stratigraphy.

expSedRamp(data, sr_start, sr_end)

Arguments

data

Input data set should consist of a matrix with 2 columns with the first being time and the second being the proxy

sr_start

Initial sedimentation rate (in cm/kyr).

sr_end

Final sedimentation rate (in cm/kyr).

Value

Returns a list which contains 10 elements element 1: time series in the depth (m) domain element 2: sedimentation rate curve

References

Routines for astrochronologic testing, astronomical time scale construction, and time series analysis <doi:10.1016/j.earscirev.2018.11.015>

Author

Based on the sedRamp function of the 'astrochron' R package.

Examples

data_1 <- etp(
tmin = 0,
tmax = 4000,
dt = 1,
eWt = 1.5,
oWt = 0.75,
pWt = 1,
esinw = T,
solution = NULL,
standardize = T,
genplot = T,
verbose = T
)
#> Error in etp(tmin = 0, tmax = 4000, dt = 1, eWt = 1.5, oWt = 0.75, pWt = 1,     esinw = T, solution = NULL, standardize = T, genplot = T,     verbose = T): could not find function "etp"
data_ramped_ls <- expSedRamp(data_1, sr_start = 1, sr_end = 5)
#> Error: object 'data_1' not found